deidentify/requirements.txt
cyclonedx-python-lib>=11.0.0,<12.0.0
